requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-pptx==0.6.23
gTTS==2.5.4
jinja2==3.1.4
numpy==1.26.4
scikit-learn==1.4.2
aiofiles==23.2.1
requests==2.32.3

reportlab==4.2.2
